docs: rename Dropdown to Select
